+++ CHANGELOG.md
@@ -3,3 +3,10 @@
 ## v0.1
 1. Region model & migration
 2. Rake task unidom:geo:china:import
+
+## v0.2
+1. Town model & migration
+2. Improved the Region model to add the #towns association
+3. Renamed the unidom:geo:china:import Rake task to unidom:geo:china:region:import
+4. Implemented the unidom:geo:china:town:import Rake task
+5. Improved the Ruby Gem Specification to depend on unidom-common v0.5